Initial fix for failure to detect/manage windows located on another screen when program starts
Turns out that "NSWorkspace sharedWorkspace runningApplications" only detects
applications/windows on the current screen, which means you can't find windows
on another screen until you actually switch to that screen for the first time.
This contributed to the code not being able to ever find/manage windows that
existed when xnomad was started, but were located on a different screen than
xnomad.

The changes here cause the program to look for new applications/windows whenever
there is a screen switch.